Moved MediatR integration into its own assembly
