+++ src/Libraries/MediatR/MediatR/PipelineBehaviors/CatchUnhandledExceptionsPipelineBehavior.cs
@@ -3,7 +3,7 @@ using System.Threading;
 using System.Threading.Tasks;
 using MediatR;
 
-namespace Dgt.CrmMicroservice.WebApi.PipelineBehaviors
+namespace Dgt.MediatR.PipelineBehaviors
 {
     public class CatchUnhandledExceptionsPipelineBehavior<TRequest, TResponse> : IPipelineBehavior<TRequest, TResponse>
         where TRequest : notnull
