+++ src/Microservices/CrmMicroservice/CrmMicroservice.WebApi/Operations/Contacts/GetContactsByNameQuery.cs
@@ -3,6 +3,7 @@ using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using Dgt.CrmMicroservice.Domain;
+using Dgt.MediatR;
 using FluentValidation;
 using MediatR;
