+++ src/dgt-mono.sln
@@ -32,6 +32,10 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CrmMicroservice.Infrastruct
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CrmMicroservice.Infrastructure.Caching.Tests", "Microservices\CrmMicroservice\CrmMicroservice.Infrastructure.Caching.Tests\CrmMicroservice.Infrastructure.Caching.Tests.csproj", "{DC3D9E6A-E610-4C5C-890A-FE6DB7C2794C}"
 EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "MediatR", "MediatR", "{5BB01E4B-2A58-424A-BDEB-522A4CEFE4FC}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MediatR", "Libraries\MediatR\MediatR\MediatR.csproj", "{A1D6D118-360E-483D-9F8C-52F1C9C58689}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -52,6 +56,8 @@ Global
 		{9DECB860-529D-4E86-B930-563347507641} = {DE7D3DBB-F3D9-47A5-9F10-B327FAC14570}
 		{E3348450-0F90-4C15-BB3A-DA99FC00AA0F} = {FA070662-C4DE-41DA-A021-1145D2C6C23B}
 		{DC3D9E6A-E610-4C5C-890A-FE6DB7C2794C} = {FA070662-C4DE-41DA-A021-1145D2C6C23B}
+		{5BB01E4B-2A58-424A-BDEB-522A4CEFE4FC} = {1162B80F-1E56-4FEB-9ECF-53D285D00F39}
+		{A1D6D118-360E-483D-9F8C-52F1C9C58689} = {5BB01E4B-2A58-424A-BDEB-522A4CEFE4FC}
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
 		{DAB1CE66-3240-4EA7-9E8D-11257F3290B6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
@@ -94,5 +100,9 @@ Global
 		{DC3D9E6A-E610-4C5C-890A-FE6DB7C2794C}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{DC3D9E6A-E610-4C5C-890A-FE6DB7C2794C}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{DC3D9E6A-E610-4C5C-890A-FE6DB7C2794C}.Release|Any CPU.Build.0 = Release|Any CPU
+		{A1D6D118-360E-483D-9F8C-52F1C9C58689}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{A1D6D118-360E-483D-9F8C-52F1C9C58689}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{A1D6D118-360E-483D-9F8C-52F1C9C58689}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{A1D6D118-360E-483D-9F8C-52F1C9C58689}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 EndGlobal
